requirements.txt
gradio==5.32.0
google-genai==1.18.0
Pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
asyncio
aiohttp==3.9.1 
pygame==2.5.2 
numpy
langchain==0.3.17
langchain-core==0.3.58
langchain-community==0.3.16
langchain-google-genai==2.1.4
pydantic-core==2.23.4
pydantic-settings==2.7.1
pydantic==2.9.2